Update building-an-admin-notice.md
This update to this tutorial page is far from perfect, but it is a giant step towards making it much more understandable to a novice Piklister. 

The {{<comment-block-styles>}} page mentioned doesn't really exist as far as I know, but it definitely should. It NEEDS to be a part of the reference documentation. 

Is there ANY way to get higher contrast for the code blocks? Grey-on-black is horribly difficult to make out.  Maybe tweak your CSS for that? Make it white or yellow on black? Letting ANYTHING (like low contrast fonts) get in the way of comprehension is a mistake.